[ADD] feed 요청 파라미처 수정

+++ src/main/java/stackpot/stackpot/repository/FeedRepository/FeedRepository.java
@@ -32,7 +32,7 @@ public interface FeedRepository extends JpaRepository<Feed, Long> {
     List<Feed> findFeeds(
             @Param("category") Category category,
             @Param("sort") String sort,
-            @Param("lastFeedId") Integer lastFeedId, // ✅ Integer로 변경
+            @Param("lastFeedId") long lastFeedId, // ✅ Integer로 변경
             @Param("lastLikeCount") long lastLikeCount, // ✅ likeCount 기준 페이징을 위한 추가 파라미터
             Pageable pageable);
     List<Feed> findByUser_Id(Long userId);
